requirements.txt
2captcha-python==1.1.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
branca==0.5.0
bs4==0.0.1
certifi==2022.9.14
charset-normalizer==2.1.1
exceptiongroup==1.0.0rc9
folium==0.12.1.post1
h11==0.14.0
html5==0.0.9
idna==3.4
Jinja2==3.1.2
logger==1.4
lxml==4.9.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
numpy==1.23.3
outcome==1.2.0
pandas==1.5.0
protobuf==3.20.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
undetected-chromedriver==3.1.6
urllib3==1.26.12
webdriver-manager==3.8.4
webencodings==0.5.1
websockets==10.3
wsproto==1.2.0
